🧹 Mason: AppLayout extraction
🎯 What: Extracted `AppHeader`, `GlobalError`, and `RetroBackground` from `AppLayout`.
💡 Why: This drastically reduces the size and complexity of `AppLayout`, improving maintainability and readability by keeping the layout component focused purely on high-level structure.
✅ Verification: Ran `pnpm lint`, `pnpm test`, and `pnpm test:e2e` to ensure type-safety and functionality are preserved.
✨ Result: `AppLayout` is now much smaller and easier to reason about.
