+++ .jules/code-health.md
@@ -1,0 +1,1 @@
+Extracted AppHeader, GlobalError, and RetroBackground from AppLayout.
